Move services configuration to separate file Add proper logging support
